+++ libs/community/langchain_community/memory/__init__.py
@@ -1,0 +1,28 @@
+"""**Memory** maintains Chain state, incorporating context from past runs.
+
+**Class hierarchy for Memory:**
+
+.. code-block::
+
+    BaseMemory --> BaseChatMemory --> <name>Memory  # Examples: ZepMemory, MotorheadMemory
+
+**Main helpers:**
+
+.. code-block::
+
+    BaseChatMessageHistory
+
+**Chat Message History** stores the chat message history in different stores.
+
+**Class hierarchy for ChatMessageHistory:**
+
+.. code-block::
+
+    BaseChatMessageHistory --> <name>ChatMessageHistory  # Example: ZepChatMessageHistory
+
+**Main helpers:**
+
+.. code-block::
+
+    AIMessage, BaseMessage, HumanMessage
+"""  # noqa: E501
